pyproject.toml
[tool.poetry]
name = "pydantic_appconfig"
version = "1.0.0"
description = "Helper package for using AWS App Config with Pydantic"
include = ["pydantic_appconfig/py.typed"]
authors = [
        "Validus Tech Team <techteam@validusrm.com>",
        "Fergus Strangways-Dixon <fergusdixon101@gmail.com>",
    ]
readme = "README.rst"
repository = "https://github.com/Validus-Risk-Management/aws-appconfig-pydantic"

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.20.8"
botocore = "^1.23.8"
PyYAML = "^6.0"
pydantic = "^1.8.2"

[tool.poetry.group.test.dependencies]
pytest = "^7.2"
pytest-mock = "^3.6.1"
freezegun = "^1.1.0"
pytest-cov = "^2.12.1"
pre-commit = "^2.13"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
flake8 = "^3.9"
isort = "^5.9"
mypy = "^0.910"
pre-commit = "^2.13"
flake8-absolute-import = "^1.0"
flake8-blind-except = "^0.2.0"
flake8-breakpoint = "^1.1.0"
flake8-bugbear = "^21.4.3"
flake8-builtins = "^1.5.3"
flake8-cognitive-complexity = "^0.1.0"
flake8-comprehensions = "^3.5.0"
flake8-docstrings = "^1.6"
flake8-eradicate = "^1.1.0"
flake8-implicit-str-concat = "^0.2.0"
flake8-mutable = "^1.2.0"
flake8-print = "^4.0.0"
flake8-pytest-style = "^1.5.0"
flake8-simplify = "^0.14.1"
pep8-naming = "^0.12.1"
types-PyYAML = "^5.4.10"
boto3-stubs = "^1.20.8"

[tool.isort]
profile = "black"

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
plugins = "pydantic.mypy"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
